Add copy and paste functionality for protected datasets
