+++ context/app/static/js/components/workspaces/NewWorkspaceDialog/NewWorkspaceDialogFromSelections.tsx
@@ -2,11 +2,15 @@ import React from 'react';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
 import MenuItem from '@mui/material/MenuItem';
+import IconButton from '@mui/material/IconButton';
+import ContentCopyIcon from '@mui/icons-material/ContentCopy';
+import InputAdornment from '@mui/material/InputAdornment';
 
 import NewWorkspaceDialog from 'js/components/workspaces/NewWorkspaceDialog';
 import WorkspaceField from 'js/components/workspaces/WorkspaceField';
 import ErrorMessages from 'js/shared-styles/alerts/ErrorMessages';
 import useProtectedDatasetsForm from 'js/components/workspaces/NewWorkspaceDialog/useProtectedDatasetsForm';
+import { useHandleCopyClick } from 'js/hooks/useCopyText';
 import { useCreateWorkspaceForm } from './useCreateWorkspaceForm';
 import { CreateWorkspaceFormTypes } from './types';
 
@@ -15,6 +19,7 @@ function NewWorkspaceDialogFromSelections() {
     useProtectedDatasetsForm();
 
   const { control, errors, setDialogIsOpen, ...rest } = useCreateWorkspaceForm({});
+  const handleCopyClick = useHandleCopyClick();
 
   return (
     <>
@@ -35,6 +40,16 @@ function NewWorkspaceDialogFromSelections() {
               label="Protected Datasets"
               value={protectedHubmapIds}
               error
+              InputProps={{
+                endAdornment: (
+                  <InputAdornment position="end">
+                    <IconButton onClick={() => handleCopyClick(protectedHubmapIds)}>
+                      <ContentCopyIcon />
+                    </IconButton>
+                  </InputAdornment>
+                ),
+                readOnly: true,
+              }}
             />
             <Button sx={{ mt: 1 }} variant="contained" color="primary" onClick={removeProtectedDatasets}>
               Remove Protected Datasets ({protectedRows.length})
